backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pdfplumber==0.10.3
openpyxl==3.1.2
pandas==2.1.4
python-dotenv==1.0.0
openai==1.3.7
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
requests>=2.31.0



